Added first test of a multitask mediator
Deals with #108. The first of several multitask mode selectors now has a
JUnit test case. The hard part of simulating a game of pacman and
checking which mode to use at each step has been done, though there is
still one incomplete test. Still, this can serve as an example of how to
create the other tests.