#280 Fix Extended Tetris Exctractor Test
Ran the JUnit Test to see where the fialure was occurring; found a
NullPointerException on line 26 of the JUnit Test, as documented in the
file. Checked the TetrisState class to see where the
NullPointerException was happening. Found that lines 81 - 101 caused the
Exception, but the test itself worked when all Parameters were set to
"true." Found that all the Parameters causing exceptions exist in the
Parameters class, and no misspellings nor incorrect capitalization
existed. Test still fails when the Parameters are not commented out, but
the source of the problem may have been found.