#126 Pinball Domain
Created a PinballViewer class and filled it with code from the
PinBallGUI and PinBallGUIReplay classes. Simplified the evaluation code
in the PinballTask class, and implemented a Visual way to evaluate the
PinballTask.
When runing the PinballTask, the console prints out, "Ball or target not
loaded in data/pinball/pinball_simple_single.cfg" when trying to
evaluate the Parents. However, the PinBallCanvas did appear for a brief
second the first time, so it may be that once this error is fixed, the
PinballViewer will work.