#126 Fixing Pinball Domain Config Files
Running the PinballTask appears to be changing the Pinball config files
in the data folder. The files should be text files describing the
polygons in the task, but running the task changes the files into
garbage text for some reason. This commit changes them back into text,
and the source of the problem will be identified soon.