#126 New Pinball Batch Files
Created a new Batch file for the PinballTask, allowing for the
pinball_hard_single.cfg file to be run as opposed to the
pinball_simple_single.cfg file. Tested both Batch files in Eclipse with
watch:true stepByStep:true, and they both function. Renamed the original
to Pinball-Simple, and confirmed it to be working.