#303 Fixed the Distance Fitness Errors
The numObjectives() method in the PinballTask class wasn't functioning
properly for the new MO code, so an If statement based on the moPinball
Parameter was added, now allowing for two Objectives when set to true.
The distance "reward" in the oneEval() method has been set to negative,
making it a punishment to have a larger ending distance from the Target.