#271 Created Amplitude.java
Created a new class and moved over the CPPN method and the methods used
to extract an amplitude array from an audio input stream. Also fixed
resulting method call errors throughout the package. This class is where
I will put the methods that save amplitude arrays as WAV and MP3 files,
and the methods that save CPPNs as WAV and MP3 files.