#308 Added Behavioral Diversity "Hooks" to PinballTask
Overrided the prep(). getBehaviorVector(), and cleanup() methods from
the NoisyLonerTask<T> class and added the appropriate data for the
PinballTask into each of them. Created a global ArrayList<Doubles> to
keep track of the X and Y coordinates of the Ball at the end of
oneEval(), and created a new Batch File called MOBDHard for the
PinballTask with the parameters needed for Behavioral Diversity.