#271 Method that saves file from input byte array
Wrote a method called saveFileFromArray that takes in an array of bytes
and saves it as a file. This works for WAV and MIDI files, but not for
mp3 files. This is because the methods that I have to convert an audio
file into a byte array are not applicable for mp3s. I'll have to write
code that can convert an mp3 file into a byte array. Also, the saved wav
file will play, but the saved midi file won't.