#271 Additional attempts to save MP3 file
I've looked through all of my available code to see if there is a way to
save an mp3 as a file from a byte array. I haven't had any success
because our JLayer implementation for playing mp3s doesn't have any of
the conversion functions I am using for wav and midi files. I had an
idea that if I figured out how to convert the MP3 file into an
AudioInputStream, I could use other methdos available to me to save the
file. I broke down the extractAmplitudeData...() methods in
SoundAmplitudeArrayManipulator so that I could return a byte array from
an input AudioInputStream, and then attempted to use the audioStream()
method in WAVUtil to convert the mp3 file to an AudioInputStream because
it looked very generalized, but I got an UnsupportedAudioFileException.
Currently trying to figure out another method that can convert an mp3
file to an AudioInputStream, although I'm still not sure this is the
best route.