#271 Method that saves file from CPPN
Wrote method that saves a file from a generated amplitude that uses a
CPPN (have not tested this). I think the method is general enough to
work for both WAVs and MP3s, but I'm not sure. To write this method, I
had to be able to convert from a double array to a byte array, so I
found a method that does so and wrote it in ArrayUtil.