#132 Created Board Games Package and Two Games
I've created a new Package in the MM-NEAT Directory called BoardGames.
This Package is currently populated with two new Interfaces (BoardGame
and BoardGameState) and two new Classes (BoardGamePlayer and
BoardGameViewer). The Viewer doesn't work yet, but the rest have bee
used to create two Sub-Packages. Package ttt has a working game of
Tic-Tac-Toe with a randomized player. RockPaperScissors is in the works.
All classes needed to build the game have been created, but it still
needs to be implemented.