More board game changes
The possibleBoardGameStates method should indicate which player is about
to play next, in order to be able to create all possible successor
states.

Also, it is bad practice to print things out from within a class,
because it may not always be used in the way that you happen to be using
it in your tests. It is better to have methods for accessing a String
representation if needed.