More big changes to board game tasks (not done yet)
#132
Made BoardGamePlayer an interface.
When a class had a default method implementation that was never used, I
changed it to be abstract instead of giving it a definition.
Simplified move() method of TTT (will change it more with next commit)
Other improvements and generalizations.

Note that RPS does not currently work, but running the tester still
works for TTT