#306 Manipulating linePlot to avoid using DrawingPanel
Had to use linePlot in my wavePlotFromDoubleArray method in
GraphicsUtil, but changed it so it used a BufferedImage to initialize
Graphics and returned a buffered image. Now there are three versions of
linePlot: the original that takes in the DrawingPanel (to avoid having
to change all areas of code that call the method) one that takes in a
Graphics instance, and one that creates and returns a BufferedImage. The
wavePlotFromDoubleArray method uses this third method and is called in
the BreedesizerTask to get the button image.