#311 Created inputs for time, sin(time) and bias
Created checkboxes for time, sin(time) and bias. Had to manipulate
amplitudeGenerator() in SoundAmplitudeArrayManipulator so that it
included inputMultipliers as a parameter. Created a copy of the original
method for this so it didn't mess up other code in
edu.utexas.cs.nn.util.sound. Now checking and unchecking the boxes
successfully manipulates the sound generated as well as the image.