#310 Save feature for breedesizer progress
Save function for breedesizer is implemented so that it saves both the
image and the sound produced locally. Was able to successfully implement
saving the images using Picbreeder code, but am having a harder time
saving the audio file. writeSingleChannel() in
SoundAmplitudeArrayManipulator was tested with an already constructed
WAV file to see whether our AudioFormat initialization was the problem.
the WAV file played and sounded the same, but seemed to be cut short.
Will continue working on the functioning of saving the audio file
tomorrow.