#313 Finished Implementing OneStepEval Player
The OneStepEval General BoardGamePlayer has now been fully implemented
as specified in the Issue. The Constructor takes a <T extends Network>
and stores that in a global variable, network. When takeAction() is
called, it stores a List of all possible BoardGameStates starting from
the given State, stores the outputs from the network, and then returns
the BoardGameState that produced the highest network output.