#310 clarifying comments/testing
Tested whether converting a WAV file to a double array and playing the
double array was the same as saving the double array using
writeSingleChannel() and playing it, because saving the file and then
playing it cut off about a second of audio. Discovered that method in
StdAudio that plays double arrays doesn't work properly, so will have to
troubleshoot that as well.