#132 TicTacToe Random Player Works Again
The "multiple turns in a row" error with the General
BoardGamePlayerRandom has been fixed. It was an error with the Copy
Constructor; using the original TicTacToeState as a Parameter would
cause the Constructor to continue to have Pointers to the old State.
With that error fixed, it may now be possible to develop and evolve the
OneStepEval TicTacToe Player.