#312 Got white keys to work in keyboard interface
Wrote out method for mouse click action in the Keyboard class. This
method plays frequency at an index in the array of white key frequencies
based on the value of the x coordinate divided by the key width. Had to
create a getter for inputMultipliers and a method in the Keyboard class
that sets the CPPN to make this function. Also had to account for the
width of the edge of the keyboard window slightly skewing the numbers
for the boundaries between notes.