#312 Black keys now work!
Modified mouseClicked() so that black keys can now play. This involved
adding -1 at regular intervals in the double array of black key
frequencies so that they would be spaced out properly, and initializing
a boolean that is set to true if there is a black key in the region, so
that white key is set as default when the boolean is false. Added extra
-1s at the end of the double array of black keys to avoid
IndexOutOfBoundException when playing the top two notes.