#329 Successfully added key listeners to keyboard interface!
Implemented KeyListener in Keyboard.java and implemented the method
keyPressed(), which plays a note frequency that corresponds to the given
index of a character pressed on the QWERTY keyboard. Also fiddled with
the method that plays the generated sounds in MiscSoundUtil so that the
sounds can be interrupted by other click actions or key presses. close
#329