Added HyperNEAT substrate info to StaticOpponentBoardGameTask
Related to issue #331 . This process needs to be repeated for the
coevolution board game tasks to, but in a way that avoids repeating
code. The meat of these methods can probably actually be moved to the
BoardGameUtil class ... just make the width and height of the board
parameters to a static method for computing the substrate information.
Then those static methods can be called from the required HyperNEAT
methods in StaticOpponentBoardGameTask and also in
SinglePopulationCompetativeCoevolutionBoardGameTask, etc.

I also moved SinglePopulationCoevolutionTask into the task package since
this is not specifically related to board games (though board games is
the first domain to use this approach).