#324 Othello Game Fully Works; Error Finding Moves
The OthelloState Class now fully works and only allows completely valid
moves to be played, updating the BoardGameState correctly. However, the
BoardGamePlayerRandom is now unable to play the game to the end; for
some reason, the possibleBoardStates() method now returns empty
ArrayLists towards the end of the game before reaching an End State.
Based on the games I watched, it appears to be unable to move over more
than two enemy Chips. I suspect that this may be causing the
possibleBoardStates to be returning fewer Moves than there are possible.
Further testing and fixing is necessary.