#310 save feature now works for images and audio
Can finally save audio generated from CPPNs in breedesizer! Did a lot of
digging into SourceDataLines when I was trying to fix playing MIDI files
with a CPPN and ended up writing a method that was meant to save an
audio file with double-to-byte array conversions that were then read
into an AudioInputStream and written to a file. This method didn't work
(was saving files but they were cut short and sounded very different
from the original CPPN-generated sound), but I found that it was very
similar to one of the save() methods in MiscSoundUtil. I used the save()
method instead and it worked perfectly! Incorporated it into the
breedesizer and have started to to a major cleanup of all sound utility
methods and the breedesizer, and am adding clarifying comments.