#331 OthelloState EndState Works
The handling for being unable to Move is now implemented; if a Player
cannot make a valid Move, they must pass their turn. The game ends if
both Players are unable to make a Move.
The capturing system needs to be changed. If more than one Chip can make
a Move to a selected Space, all Enemy Chips from both Moves are
converted, not just the Chips from the one Selected Move.