#316 Updates on remixing sound
Changed Remixbreeder so it extends BreedesizerTask instead of
InteractiveEvolutionTask because the code was so similar. Also added in
double array representation of a wav file, but cannot convert from an
input wav file to a double array at the moment - read() method from
StdAudio isn't functional. Currently trying to figure out conversion
from a wav file to a double array