#316 Clarifying comments and simplification of readByte() method
Added header comment to amplitudeRemixer and simplified readByte()
method in the hopes that it would cause the read() method to be
functional so I could convert the wav file input to Remixbreeder to a
double array that can be played. Unfortunately, this didn't work. Still
looking for ways to convert a WAV file to a double array