#324 Updated Othello System
The OthelloState has been updated in accordance with the comments left
yesterday. However, the move() method is still having Errors; this time,
the Players will never make a Move to the far Left Column (y == 0).
There are several comments throughout the move() method detailing the
nature of the Error, but it definitely not a problem with the
possibleBoardStates() method. Any help on the issue would be greatly
appreciated.