#316 More attempts to convert WAV file to double array
Methods I've tried:
-converting directly from a string reference to a file to a
corresponding double array (read())
-2 different approaches to converting an array of bytes to an array of
doubles (extractDoubleArrayFromAmplitudeByteArray() and toDoubleArray()
in SoundToArray.java)
-Extracting an int array from an audio input stream representation of
the wav file and casting the content at each index to a double
(extractAmplitudeDataFromAmplitudeByteArray())
None of these methods work. I'm wondering if there's a way to remix the
amplitude without using a double array in the Remixbreeder, but that may
not be possible. Currently pretty stuck on this task