#324 Minor EndState Change to Othello
The Number of Passes is reset to 0 if a Player makes a Move. The reason
for this is to prevent the Game from ending early; if a Player passes
early in the Game, the next time either Player passes will be the
EndState. With this update, the Game only ends if both Players are
unable to make a Move.