Interactive Evolution interface generalization and improvement
I moved more content from the individual evolution programs (picbreeder,
breedesizer, and remixbreeder) into InteractiveEvolutionTask. Handling
of input disabling checkboxes is now completely generalized.

For sound evolving systems, I made is so that unselecting a sound stops
the playback.

For RemixBreeder, I added a button to play the original sound, which now
no longer plays on program start.

This adds features to #316 and some functionality to #271