#340 Tried to adjust format of files directly
Attempted to get sound to play for all formats of files by adjusting the
format of input files so they were restricted to 16 bit, mono, 2
bytes/frame, and PCM_SIGNED. Did this by creating a mehtod that took in
the original AudioFormat and adjusted it according to these
specifications. All files are now playable, but don't sound like what
they should.