#330 Enabled button for MIDI playback in Breedesizer
Created button that plays a MIDI file using the most recently clicked
wave in the Breedesizer as the "instrument". Had to override some of the
code in the constructor/respondToClick methods so it wouldn't show up in
the Remixbreeder because RemixbreederTask extends BreedesizerTask. The
button is still fully functional, but I stil have to fix MIDI playback.