#330 MIDI playback now mostly functional!
Fiddled around a lot with conversion between ticks (time measurement of
events in MIDI files) and milliseconds, and changed the freqFromMIDI
method so that it took in a single track and if the note was being
turned off, it set the duration of the note to the difference between
the starting note and ending note. This value only works if we multiply
it by 10 for some reason. Right now, the MIDI playback doesn't work for
chords - it only plays the lowest note if multiple notes are being
played simultaneously. Also, as far as we know, it only works for songs
with only one instrument. Right now, our code is specified to only work
with a single track, but multiple tracks = multiple instruments.