#339 JUnit Test Cases for TicTacToe and Othello
The JUnit Test Cases for TicTacToeState and OthelloState have been
completed. Every test works except for the possibleBoardStates() method
for the OthelloState, but I cannot identify why the test fails. The test
is strutured similarly to the TicTacToe test, which passes. Any insight
on this would be appreciated.
Otherwise, the Test Cases work perfectly. The possibleBoardStates()
method has been modified to return a Set<BoardGameState> instead of a
List, and the appropriate changes have been made in all BoardGameState
classes and BoardGamePlayer classes.