Provided a way to interrupt "MIDI" playback in Breedesizer
#330  Added an inner class to MIDIUtil that controls playback of note
sequences from a CPPN. This is basically a Thread that contains the
previously written playback loop, but it also has a method that can be
called to stop playback.

This new class is used in Breedesizer so that any other button click
will stop the playback.

The Thread trick used here is similar to what is done already when
playing individual amplitude arrays. However, that earlier code is
problematic because it uses global static variables and does not return
a Thread that controls cancellation of the playback. The previously
existing sound playing approach should be modified to be more like this
new approach to sound playback interruption.