Figure out how to track score history
There are some simple comment cleanups in this commit, but I also found
old code for tracking the history of scores achieved by every agent.
Though the ScoreHistory class seems complete, it is not used anywhere. I
recall considering the use of all old scores in the past, but never
finding a way of doing it. However, now that I've found this code, I can
probably find a way to use it to accomplish #334 .This is not
implemented yet, but I've left some comments to guide me on how to do
it.