#340 further generalization of WAV file playback
Removed all references to static final BYTES_PER_SAMPLE because it
applied exclusively to the AudioFormat of 16 bit files. This caused an
IndexOutOfBoundException when I tried to convert an 8 bit file to a
double array, so I'm wondering if there is a way to generalize the
defined sample buffer size. Will work on moving code that plays double
arrays into thread class tomorrow so that line instances don't have to
be static - will make fixing issues with 8 bit files easier.