Found issue with saving files in Remixbreeder
Found that there was an issue with saving audio files in the
Remixbreeder because it extended the save method of the Breedesizer, and
the number of inputs in the Breedesizer and Remixbreeder don't match.
Made an additional protected method in Breedesizer called saveSound()
that includes the method call to save the file from a CPPN. Overrode
this in Remixbreeder and called a new method I wrote called
saveRemixedFileFromCPPN that is the same as saveFileFromCPPN except that
it uses amplitudeRemixer instead of amplitudeGenerator. I also removed
an AudioFormat parameter from several saving methods because I saw that
the AudioFormat was never used. Now the sounds aren't playing back
correctly when they are saved in the Remixbreeder (it saves and looks
similar, but plays as if it were sped up), but Breedesizer save function
is still completely functional.