#344 Fixed saving issues with Remixbreeder!
It turns out that the AudioFormat was a necessary parameter to save a
file from a double array, but it wasn't being used. A very specific
AudioFormat was being initialized and used in saveFromArray(), so I
created a copy of the method and put in the AudioFormat as the
parameter. In the Remixbreeder, I accessed the AudioFormat of the input
WAV file in the method call, and now playback of saved files is
identical to original playback. close #344