#330 Moved large amount of code in PlayDoubleArray to an inner class
Moved several methods and variables to a private inner class called
AmplitudeArrayPlayer that extends Thread in PlayDoubleArray. This was
done to get rid of the initializers and make the SourceDataLine
non-static so that multiple sounds can potentially be played back at
once. This broke a couple of things in the code - firstly, playback can
no longer be interrupted on the interface. Secondly, WAV file playback
is incorrect again because code can no longer call changeAudioFormat()
to access the input WAV file's specific audio format. Will need to fix
this as soon as possible. There is a backup version of the jar file
saved just in case we cannot solve this issue before Mad Science Monday.