#330 Fixed interruption failure
Fixed interruption failure in Breedesizer and Remixbreeder by making
PlayDoubleArray methods return an AmplitudeArrayPlayer instance, so
clicking can be determined based on the state of the arrayPlayer.
Booleans to allow interruption were also set as parameters for several
methods involved with playing double arrays. Now sounds can overlap if
we so choose (we turned this off for Breedesizer and Remixbreeder) and
they can be stopped if a button is clicked a second time.