#330 Fixed AudioFormat for original file playback
AudioFormat for original file playback just needed to be added as a
parameter when original sound was played to play back the right audio.
However, when this was enabled, if the PlayOriginal button was clicked a
secondary time while it was playing the original file, it would just
start it over and overlap rather than stopping the sound. Had to add a
method called isPlaying() to AmplitudeArrayPlayer inner class that
returns whether the sound is playing or not, and include that as a
condition in the if statement that stops playback. The if statement that
played the original file became an else if so that the two actions were
seperated.