#331 BoardGame HyperNEAT Prep Work
The Players for BoardGameTasks have been generalized in
StaticOpponentBoardGameTask. and SinglePopulationCompetativeCoevolution.
The Parameters "opponentHeuristic," "player," and "playerHeuristic" have
been added to the Parameters Class. A blank Constructor has been added
to the NNBoardGameHeuristic to prevent the Tasks from crashing at
startup; the NNBoardGameHeuristic is re-created later in both Tasks with
the correct Network to ensure proper functionality. In addition, a
setHeuristic() method has been added to all existing BoardGamePlayers.
The currently existing Batch Files are being tested now. As of this
moment, they appear to be functioning properly. The Human Players have
been removed from the Othello, RPS, and TicTacToe BoardGames; a combined
2DHumanPlayer class may be created later if so desired.