#340 Attempts to play 8-bit files
Generalized code that plays audio so that it would use the maximum value
for 8-bit files if it was an 8 bit file and vice versa for a 16 bit file
(it was only using the maximum for 16 bit files before). Also tested to
see if byte array conversion in SoundToArray and the lack of
generalization in the read() method of SoundtoArray were issues.
Changing these things are steps in the right direction, but 8-bit WAV
files still cannot play.