#319 Started process of evolving animations
Created new class called AnimationBreederTask that extends
InteractiveEvolutionTask. Intended to create an interface similar to
Picbreeder, but with animations instead of images. Also created an
additional utility class meant to animate an array of images by
extending JApplet. I found online code that is meant to do this and
copied it in, but it will most likely have to be adjusted.