#322 BoardGameFeatureExtractors Created
The BoardGameFeatureExtractor Interface has been created, along with the
RawBoardGameFeatureExtractor and SplitPiecesBoardGameExtractor. The
former returns the double Array produced by the getDescription() method
in BoardGameState. The latter returns a double Array with a length of
numberOfPlayers*getDescription().length. Each segment of
getDescription.length contains every Empty Space on the Board and the
Pieces of one unspecific Player, but this can be changed later.