Fixed assertion
Assertion in NSGA 2 was comparing the wrong genotyp IDs. The assertion
was never an issue when running with assertions off, but I don't think
there was an actual error. However, it is worth noting that the genotype
returned by crossover can (will?) have the same ID as its input. This is
why it is important to copy the genotype before applying crossover,
since the copy operation does result in a new ID.