#330 New code that separates MIDI data into lines
Created code in MIDIUtil that is intended to separate MIDI file data
into singular arrays representing different voices, so that multiple
voices can be played at once. The arrays that resulted were incredibly
long (hundreds of thousands of indexes), so we also created a way to
extract the frequencies and lengths from the lists to take out
unnecessary data. This prints individual voices successfully, but will
not work with the inner class in MIDIUtil that plays sounds because it
uses an array list of doubles and the inner class uses double arrays.