#330 Removed popping noise w/MIDI playback
Wrote some tests to see if synchronized playback with MIDI files would
work with new code. Converted array lists to double arrays so they were
compatible with methods playing MIDI from CPPN, but sound does not play
in a synchronized fashion and there was a strange popping sound before
each note. Fixed this issue by creating a method called removePops()
that gradually increases and decreases the volume at the beginning and
ending of each note to avoid sudden drastic changes in volume. This
worked for the most part, but still need to figure out how to
synchronize playback.