#330 Saving notes + lengths into larger array
In an attempt to synchronize MIDI playback, created a method called
lineToAmplitudeArray() that takes in the data from a soundLine and reads
it to a double array corresponding to all of the notes with a number of
indexes per note that corresponds to the note lengths. Haven't figured
out playback for multiple voices, but sound is much cleaner now and the
popping noise is gone.