#330 MIDI playback works! (for fur elise)
Changed inner class of MIDIUtil so that it took in an array list of
double arrays rather than two individual double arrays representing the
length and frequency. Looping through a double array representation of
the amplitudeArrays allowed playback to mostly be successful! All of the
voices are playing in sync, but they become more and more out of sync as
time progresses. Tried to play a file with multiple tracks, but this was
not successful.  I think there are some special cases associated with
MIDI file formatting that haven't been accounted for in the code