#350 Created BehavioralDiversity for BoardGames
Created a method in the BoardGame Interface called getDescription();
it's currently used to return the BoardGameState's descriptor. This
change was necessary to allow the StaticOpponentBoardGameTask to access
the BoardGameState's descriptor for Behavioral Diversity. Created a MOBD
Othello Batch File with a Minimax Piece Differential Static Opponent.
Will run the Batch File and document results.