Note these changes and add comments
@alicequint
I moved the evaluate function into the parent class of these two classes
because the code was the same in each case.

I also fixed your normalize method, which was not actually normalizing
anything (the original array was not changed). There are still some
issues with this code though. See my suggestions.

In general, you need to maintain better code comments.