Fixed network display with microRTS map sequence
Addresses issues #348 .
Now the network displays at the correct size before the evaluation
happens. This was accomplished by adding a new method to the LonerTask
that is executed BEFORE the evaluation happens. This is required,
because LonerTask displays the network before it executes the evaluation
code, but in the old code, the map size change occurred inside of the
evaluation code. Now, the map change happens in the new preEval method
(from LonerTask, but overridden in MicroRTSTask).

@alicequint
There is a very important followup issue associated with this. Similar
code changes need to be made to the coevolution code. Specifically, the
SinglePopulationCoevolutionTask needs a preEval method whose default
contents are empty, but the method can be overridden in
SinglePopulationCompetativeCoevolutionMicroRTSTask. Note that the
evaluateAll method of SinglePopulationCoevolutionTask will need to call
preEval right before evaluateGroup gets called.