#330 Trying to combine two audio files
I was looking into combining two sound waves and this stack overflow
link
(https://stackoverflow.com/questions/18453564/java-record-mix-two-audio-streams)
had an approved solution that said that the values at each index of two
byte arrays just had to be added up and divided by two if necessary. I
had already tested this with an ArrayUtil method called overlap(), but I
hadn't divided the values by two, so I went ahead and did that in this
test. It was unsuccessful - sound will play, but it sounds like the same
crackling static for each of the examples I tested