#354 NNBoardGamePlayer EndState Override Parameter
Added "heuristicOverrideTerminalStates" as a boolena Parameter in the
Parameters Class, defaulted to false. If set to true, the
NNBoardGamePlayer's evaluation of End States is overriden, returning a 1
for an undisputed win, a 0 for a tie which includes the current Player,
or a -1 for an undisputed loss or a tie which does not include the
current Player.
A getCurrentPlayer() method has been added to the BoardGameState
Interface and 2DBoardGameState Class to aid with this evaluation.