#322 BoardGameFeatureExtractor Edits
Created a separate FeatureExtractor Package within the BoardGame
Package, and moved the existing FeatureExtractor Interface and related
Classes to the new Package. Moved the getFeatureLabels() method from the
BoardGame Interface to the FeatureExtractor Interface. Changed the Raw
FeatureExtractor to only accept TwoDimensionalBoardGames, and moved the
getFeatureLabels() method from the TwoDimensionalBoardGame Class to the
Raw FeatureExtractor.
Created a FeatureExtractor class Parameter in the Parameters Class, and
changed NNBoardGameHeuristic and the BoardGameTasks to accommodate this
change. Removed TwoDimensionalBoardGameTask because it was redundant and
never called; it was made up of copy-pasted code from the other Tasks
with no extra functions added.