#355 Started creating picture remixer
Initialized class that remixes imput images. Put it in the same package
as RemixBreeder and renamed RemixBreeder task to sound remix. Not sure
if this is the best way to generalize them or if they should each go in
separate packages, though. Confused about the specifications of loading
an image's data (height, width, hue, saturation, value)