#353 Created Alpha-Beta Pruning Minimax Player
Created an Alpha-Beta Pruning Minimax Player that extends the original
Minimax Player. Added the "alpha" and "beta" parameters to the original
minimax() method, which is ignored in the original Minimax Player. The
Alpha-Beta Minimax Player uses the "alpha" and "beta" parameters in its
overriding minimax() method. Should be functioning properly.