#355 Picture Remix now functional!
Fixed some issues with extracting the HSB from the original image and
remixing it in GraphicsUtil (there was a mismatch with the number of
inputs to the CPPN) and now Picture Remix is functional. The colors of
the input images can be manipulated in super cool ways, but the
underlying image remains the same (the color mutationss are kind of like
a filter for the image). May want to look into ways to change the line
structure/stretch out underlying structure for current image in the
future. This would also be useful for remixing sounds. close #355