#359 Added remixImageWindow to Picture Remix
Added parameter called remixImageWindow to the Picture Remix interface
that specifies a small window surrounding a pixel that is manipulated
based on the HSB data to produce more interesting images. Made a nested
loop within the nested loop in remixedImageFromCPPN() to do this
(inefficient!) but it produces some very interesting images.