#359 Tried to randomize pixel outputs
Rewrote remixedImageFromCPPN so that it takes in the input pixels and
outputs them at a completely different location in the remixed image
(this required us to change the number of outputs so that the outputs
also included x-y coordinates. This randomized the code too much and the
original image is not recognizable even on the first generation, making
it no different from the original Picbreeder.