#330 Switched over to MIDI play by adding sound waves
This version adds the soundwaves together and then plays them back.
Every source that I can find on the internet indicates that combining
sounds simply boilds down to adding the waves together, so I think that
any issues we are having must be related to the volume scaling or other
steps of the conversion process. Let's move forward with this version of
the code.