#330 Many TODOs on the MIDI issue
I added many TODO comments related to the MIDI issue. However, halfway
through, I realized that I had slightly misunderstood the data
representations used at some steps of the process, making some of the
comments invalid. I went back and fixed some, but not all. The general
sentiment is the same though: make everything more efficeint, and try to
figure out what is going on in some of the tricky parts of the code.

@twerisa