#330 Removed unused code
Removed methods that weren't being used for MIDI file decoding and
experimented with ticks to millisecond conversion because the original
method was not functional. Found a calculation that is basically
equivalent to the magic number 50. Takes the default audio format's
sample rate and divides it by the microseconds per tick (got 49.3 for
Fur Elise).  Still need to fix problems with massive amount of memory
usage and remove unnecessary middle steps with the ArrayList