#367 Checkers Batch File Error Fixed
The error causing hte Checkers Batch Files to not run has been fixed;
there was an issue with the CheckersState endState() method that was
causing it to return false when it was impossible to make a Move. May
consider implementing a Pass move similar to Othello. Looked online for
official rules, but nothing was conclusive.