#330 Increasing efficiency of conversion process
Creating duplicates of methods to account for soundLines() returning a
Triple corresponding to the frequencies, lengths, and start times of a
given track. Also made a new method in StatisticsUtilities that computes
the sum of a collection of doubles for the duplicate
lineToAmplitudeArray() method. Still confused about incorporating third
part of Triple into lineToAmplitudeArray()