#357 Edited FitnessFunction Interface and Classes
The FitnessFunction Classes now update the Fitness of a Player over the
coarse of the game rather than just at the end, as was previously
implemented. During a game inside BoardGameUtil, the Fitness is stored
in an Array of Doubles with a length equal to the number of Players.
This means that the Fitness of each Player can be stored and updated
over the coarse of a game. The final Fitness is handled as the previous
implementation handled it.