Removed long unused commented code
There was some code in here dealing with the display of hierarchical
networks, but I haven't used those in a long time, and I'm honestly not
even sure that they still work (or that they ever worked). So I removed
the commented code.

This was done partly in an effort to clean up this code in advance of
having Darwin work on issue #368