New MIDI playback problem fixed, but still work to be done
#330
@twerisa
The problem was that in lineToAmplitudeArray we were not multiplying
lengths and ticks by the amplitudeLengthMultiplier. Adding this
multiplication stretches out the sounds to the correct scale. However, I
worry about rounding errors when your proper time scaling approach
computes the amplitude multiplier. We will probably have to round up to
the nearest whole number to make everything work.

This commit has a lot of commented out debugging code that I used to
generate some useful files that I will put in the google drive
artifacts.