#330 Code cleanup/testing
Commented out unused methods that will be deleted after this commit.
Also tested more playback/printing of files with multiple tracks. Every
time I run the code, it gives a NullPointerException pointing to the
first line of the case in which a note is turned off. I thought this
could be because some MIDI files have a note turned off as their first
data instance, so I created a boolean began to account for the note
being turned off not being the very first line. This didn't work either