#366 Partially Implemented 2DHumanBoardGamePlayer
Added several methods to the BoardGamePlayerHuman2DBoard class. The
class will allow a Human to play a TwoDimensionalBoardGame by clicking
on the pieces to be moved and the space to be moved to.
The getInput method handles the majority of the move handling; it will
take the Player's inputs and move a copy of the current BoardGameState,
returning the copy.
The validity of the Move is checked by takeAction() before the final
return of the Move.